refactor: move project-index fs helpers into semantic package
Relocated listDirectory/getFileMtime from src/shared/utils/fs.js to src/semantic/src/project-index/fs-helpers.js because they only support the project index traversal flows and no longer need to live in the shared CLI-facing fs utilities.
