+++ src/semantic/src/project-index/built-in-identifiers.js
@@ -2,13 +2,13 @@ import {
     parseJsonWithContext,
     normalizeIdentifierMetadataEntries,
     areNumbersApproximatelyEqual,
-    getFileMtime,
     isPlainObject
 } from "../../../shared/index.js";
 import { GML_IDENTIFIER_METADATA_PATH } from "../resources/bundled-resources.js";
 
 import { defaultFsFacade } from "./fs-facade.js";
 import { createProjectIndexAbortGuard } from "./abort-guard.js";
+import { getFileMtime } from "./fs-helpers.js";
 
 const GML_IDENTIFIER_FILE_PATH = GML_IDENTIFIER_METADATA_PATH;
