+++ src/semantic/src/project-index/project-root.js
@@ -1,15 +1,13 @@
 import path from "node:path";
 
-import {
-    walkAncestorDirectories,
-    listDirectory
-} from "../../../shared/index.js";
+import { walkAncestorDirectories } from "../../../shared/index.js";
 import {
     PROJECT_ROOT_DISCOVERY_ABORT_MESSAGE,
     createProjectIndexAbortGuard
 } from "./abort-guard.js";
 import { isProjectManifestPath } from "./constants.js";
 import { defaultFsFacade } from "./fs-facade.js";
+import { listDirectory } from "./fs-helpers.js";
 
 export async function findProjectRoot(options, fsFacade = defaultFsFacade) {
     const filepath = options?.filepath;
